pyproject.toml
[tool.poetry]
name = "sync-focus"
version = "0.1.0"
description = ""
authors = ["meow"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "1.0.0"
uvicorn = "0.23.1"
fastapi = "0.100.1"
sqlalchemy = "2.0.26"
asyncpg = "0.28.0"
greenlet = "3.0.3"
python-multipart = "0.0.6"
bcrypt = "4.0.1"
passlib = "1.7.4"
python-jose = "3.3.0"
ruff = "^0.1.5"
backoff = "2.2.1"
blinker = "1.7.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
cython = "3.0.8"
deprecated = "1.2.14"
flask = "2.3.3"
flask-sqlalchemy = "3.1.1"
googleapis-common-protos = "1.62.0"
grpcio = "1.60.1"
idna = "3.6"
importlib-metadata = "6.11.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.3"
mako = "1.3.2"
markupsafe = "2.1.5"
opentelemetry-api = "1.22.0"
opentelemetry-distro = "0.43b0"
opentelemetry-exporter-otlp = "1.22.0"
opentelemetry-exporter-otlp-proto-common = "1.22.0"
opentelemetry-exporter-otlp-proto-grpc = "1.22.0"
opentelemetry-exporter-otlp-proto-http = "1.22.0"
opentelemetry-instrumentation = "0.43b0"
opentelemetry-instrumentation-aws-lambda = "0.43b0"
opentelemetry-instrumentation-dbapi = "0.43b0"
opentelemetry-instrumentation-flask = "0.43b0"
opentelemetry-instrumentation-jinja2 = "0.43b0"
opentelemetry-instrumentation-logging = "0.43b0"
opentelemetry-instrumentation-sqlite3 = "0.43b0"
opentelemetry-instrumentation-urllib = "0.43b0"
opentelemetry-instrumentation-wsgi = "0.43b0"
opentelemetry-propagator-aws-xray = "1.0.1"
opentelemetry-proto = "1.22.0"
opentelemetry-sdk = "1.22.0"
opentelemetry-semantic-conventions = "0.43b0"
opentelemetry-util-http = "0.43b0"
packaging = "23.2"
protobuf = "4.25.2"
requests = "2.31.0"
typing-extensions = "4.9.0"
urllib3 = "2.2.0"
werkzeug = "2.3.8"
wrapt = "1.16.0"
zipp = "3.17.0"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
